Reorganise CNY2023 entities
