+++ app/main.js
@@ -50,33 +50,33 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 
 /***/ }),
 
-/***/ "./src/avo/entity/types/boost-pad.js":
-/*!*******************************************!*\
-  !*** ./src/avo/entity/types/boost-pad.js ***!
-  \*******************************************/
+/***/ "./src/avo/entity/types/cny2023/boost-pad.js":
+/*!***************************************************!*\
+  !*** ./src/avo/entity/types/cny2023/boost-pad.js ***!
+  \***************************************************/
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ BoostPad)\n/* harmony export */ });\n/* harmony import */ var _avo_entity__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/entity */ \"./src/avo/entity/index.js\");\n/* harmony import */ var _avo_constants__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/constants */ \"./src/avo/constants.js\");\nfunction _typeof(obj) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (obj) { return typeof obj; } : function (obj) { return obj && \"function\" == typeof Symbol && obj.constructor === Symbol && obj !== Symbol.prototype ? \"symbol\" : typeof obj; }, _typeof(obj); }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\n\nfunction _get() { if (typeof Reflect !== \"undefined\" && Reflect.get) { _get = Reflect.get; } else { _get = function _get(target, property, receiver) { var base = _superPropBase(target, property); if (!base) return; var desc = Object.getOwnPropertyDescriptor(base, property); if (desc.get) { return desc.get.call(arguments.length < 3 ? target : receiver); } return desc.value; }; } return _get.apply(this, arguments); }\n\nfunction _superPropBase(object, property) { while (!Object.prototype.hasOwnProperty.call(object, property)) { object = _getPrototypeOf(object); if (object === null) break; } return object; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); Object.defineProperty(subClass, \"prototype\", { writable: false }); if (superClass) _setPrototypeOf(subClass, superClass); }\n\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\n\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\n\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } else if (call !== void 0) { throw new TypeError(\"Derived constructors may only return object or undefined\"); } return _assertThisInitialized(self); }\n\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\n\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Boolean.prototype.valueOf.call(Reflect.construct(Boolean, [], function () {})); return true; } catch (e) { return false; } }\n\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\n\nvar DEFAULT_BOOST_POWER = 32;\n\nvar BoostPad = /*#__PURE__*/function (_Entity) {\n  _inherits(BoostPad, _Entity);\n\n  var _super = _createSuper(BoostPad);\n\n  function BoostPad(app) {\n    var _this;\n\n    var col = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : 0;\n    var row = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : 0;\n    var width = arguments.length > 3 && arguments[3] !== undefined ? arguments[3] : 1;\n    var height = arguments.length > 4 && arguments[4] !== undefined ? arguments[4] : 1;\n    var cutCorner = arguments.length > 5 && arguments[5] !== undefined ? arguments[5] : false;\n    var boostPower = arguments.length > 6 && arguments[6] !== undefined ? arguments[6] : DEFAULT_BOOST_POWER;\n\n    _classCallCheck(this, BoostPad);\n\n    _this = _super.call(this, app);\n    _this._type = 'boost-pad';\n    _this.colour = '#c0a040';\n    _this.solid = false;\n    _this.movable = false;\n    _this.x = col * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE;\n    _this.y = row * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE;\n    _this.shape = _avo_constants__WEBPACK_IMPORTED_MODULE_1__.SHAPES.POLYGON;\n    _this.shapePolygonPath = [];\n    if (cutCorner !== 'nw') _this.shapePolygonPath.push(0, 0);\n    if (cutCorner !== 'ne') _this.shapePolygonPath.push(width * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE, 0);\n    if (cutCorner !== 'se') _this.shapePolygonPath.push(width * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE, height * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE);\n    if (cutCorner !== 'sw') _this.shapePolygonPath.push(0, height * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE);\n    _this.boostPower = boostPower;\n    return _this;\n  }\n\n  _createClass(BoostPad, [{\n    key: \"onCollision\",\n    value: function onCollision(target, collisionCorrection) {\n      _get(_getPrototypeOf(BoostPad.prototype), \"onCollision\", this).call(this, target, collisionCorrection);\n\n      target.pushY = -this.boostPower;\n    }\n  }]);\n\n  return BoostPad;\n}(_avo_entity__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n\n\n\n//# sourceURL=webpack://cny2023/./src/avo/entity/types/boost-pad.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ BoostPad)\n/* harmony export */ });\n/* harmony import */ var _avo_entity__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/entity */ \"./src/avo/entity/index.js\");\n/* harmony import */ var _avo_constants__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/constants */ \"./src/avo/constants.js\");\nfunction _typeof(obj) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (obj) { return typeof obj; } : function (obj) { return obj && \"function\" == typeof Symbol && obj.constructor === Symbol && obj !== Symbol.prototype ? \"symbol\" : typeof obj; }, _typeof(obj); }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\n\nfunction _get() { if (typeof Reflect !== \"undefined\" && Reflect.get) { _get = Reflect.get; } else { _get = function _get(target, property, receiver) { var base = _superPropBase(target, property); if (!base) return; var desc = Object.getOwnPropertyDescriptor(base, property); if (desc.get) { return desc.get.call(arguments.length < 3 ? target : receiver); } return desc.value; }; } return _get.apply(this, arguments); }\n\nfunction _superPropBase(object, property) { while (!Object.prototype.hasOwnProperty.call(object, property)) { object = _getPrototypeOf(object); if (object === null) break; } return object; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); Object.defineProperty(subClass, \"prototype\", { writable: false }); if (superClass) _setPrototypeOf(subClass, superClass); }\n\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\n\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\n\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } else if (call !== void 0) { throw new TypeError(\"Derived constructors may only return object or undefined\"); } return _assertThisInitialized(self); }\n\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\n\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Boolean.prototype.valueOf.call(Reflect.construct(Boolean, [], function () {})); return true; } catch (e) { return false; } }\n\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\n\nvar DEFAULT_BOOST_POWER = 32;\n\nvar BoostPad = /*#__PURE__*/function (_Entity) {\n  _inherits(BoostPad, _Entity);\n\n  var _super = _createSuper(BoostPad);\n\n  function BoostPad(app) {\n    var _this;\n\n    var col = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : 0;\n    var row = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : 0;\n    var width = arguments.length > 3 && arguments[3] !== undefined ? arguments[3] : 1;\n    var height = arguments.length > 4 && arguments[4] !== undefined ? arguments[4] : 1;\n    var cutCorner = arguments.length > 5 && arguments[5] !== undefined ? arguments[5] : false;\n    var boostPower = arguments.length > 6 && arguments[6] !== undefined ? arguments[6] : DEFAULT_BOOST_POWER;\n\n    _classCallCheck(this, BoostPad);\n\n    _this = _super.call(this, app);\n    _this._type = 'boost-pad';\n    _this.colour = '#c0a040';\n    _this.solid = false;\n    _this.movable = false;\n    _this.x = col * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE;\n    _this.y = row * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE;\n    _this.shape = _avo_constants__WEBPACK_IMPORTED_MODULE_1__.SHAPES.POLYGON;\n    _this.shapePolygonPath = [];\n    if (cutCorner !== 'nw') _this.shapePolygonPath.push(0, 0);\n    if (cutCorner !== 'ne') _this.shapePolygonPath.push(width * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE, 0);\n    if (cutCorner !== 'se') _this.shapePolygonPath.push(width * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE, height * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE);\n    if (cutCorner !== 'sw') _this.shapePolygonPath.push(0, height * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE);\n    _this.boostPower = boostPower;\n    return _this;\n  }\n\n  _createClass(BoostPad, [{\n    key: \"onCollision\",\n    value: function onCollision(target, collisionCorrection) {\n      _get(_getPrototypeOf(BoostPad.prototype), \"onCollision\", this).call(this, target, collisionCorrection);\n\n      target.pushY = -this.boostPower;\n    }\n  }]);\n\n  return BoostPad;\n}(_avo_entity__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n\n\n\n//# sourceURL=webpack://cny2023/./src/avo/entity/types/cny2023/boost-pad.js?");
 
 /***/ }),
 
-/***/ "./src/avo/entity/types/ground.js":
-/*!****************************************!*\
-  !*** ./src/avo/entity/types/ground.js ***!
-  \****************************************/
+/***/ "./src/avo/entity/types/cny2023/ground.js":
+/*!************************************************!*\
+  !*** ./src/avo/entity/types/cny2023/ground.js ***!
+  \************************************************/
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ Ground)\n/* harmony export */ });\n/* harmony import */ var _avo_entity__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/entity */ \"./src/avo/entity/index.js\");\n/* harmony import */ var _avo_constants__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/constants */ \"./src/avo/constants.js\");\nfunction _typeof(obj) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (obj) { return typeof obj; } : function (obj) { return obj && \"function\" == typeof Symbol && obj.constructor === Symbol && obj !== Symbol.prototype ? \"symbol\" : typeof obj; }, _typeof(obj); }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\n\nfunction _get() { if (typeof Reflect !== \"undefined\" && Reflect.get) { _get = Reflect.get; } else { _get = function _get(target, property, receiver) { var base = _superPropBase(target, property); if (!base) return; var desc = Object.getOwnPropertyDescriptor(base, property); if (desc.get) { return desc.get.call(arguments.length < 3 ? target : receiver); } return desc.value; }; } return _get.apply(this, arguments); }\n\nfunction _superPropBase(object, property) { while (!Object.prototype.hasOwnProperty.call(object, property)) { object = _getPrototypeOf(object); if (object === null) break; } return object; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); Object.defineProperty(subClass, \"prototype\", { writable: false }); if (superClass) _setPrototypeOf(subClass, superClass); }\n\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\n\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\n\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } else if (call !== void 0) { throw new TypeError(\"Derived constructors may only return object or undefined\"); } return _assertThisInitialized(self); }\n\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\n\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Boolean.prototype.valueOf.call(Reflect.construct(Boolean, [], function () {})); return true; } catch (e) { return false; } }\n\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\n\n\nvar Ground = /*#__PURE__*/function (_Entity) {\n  _inherits(Ground, _Entity);\n\n  var _super = _createSuper(Ground);\n\n  function Ground(app) {\n    var _this;\n\n    var col = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : 0;\n    var row = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : 0;\n    var width = arguments.length > 3 && arguments[3] !== undefined ? arguments[3] : 1;\n    var height = arguments.length > 4 && arguments[4] !== undefined ? arguments[4] : 1;\n    var cutCorner = arguments.length > 5 && arguments[5] !== undefined ? arguments[5] : false;\n\n    _classCallCheck(this, Ground);\n\n    _this = _super.call(this, app);\n    _this._type = 'wall';\n    _this.colour = '#888';\n    _this.solid = true;\n    _this.movable = false;\n    _this.x = col * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE;\n    _this.y = row * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE;\n    _this.shape = _avo_constants__WEBPACK_IMPORTED_MODULE_1__.SHAPES.POLYGON;\n    _this.shapePolygonPath = [];\n    if (cutCorner !== 'nw') _this.shapePolygonPath.push(0, 0);\n    if (cutCorner !== 'ne') _this.shapePolygonPath.push(width * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE, 0);\n    if (cutCorner !== 'se') _this.shapePolygonPath.push(width * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE, height * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE);\n    if (cutCorner !== 'sw') _this.shapePolygonPath.push(0, height * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE);\n    return _this;\n  }\n\n  _createClass(Ground, [{\n    key: \"onCollision\",\n    value: function onCollision(target, collisionCorrection) {\n      _get(_getPrototypeOf(Ground.prototype), \"onCollision\", this).call(this, target, collisionCorrection);\n\n      var app = this._app;\n      var controls = app.rules['cny2023-controls'];\n      var goals = app.rules['cny2023-goals'];\n      if (!controls || !goals) return;\n      if (target === app.hero) goals.triggerLoseScreen();\n    }\n  }]);\n\n  return Ground;\n}(_avo_entity__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n\n\n\n//# sourceURL=webpack://cny2023/./src/avo/entity/types/ground.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ Ground)\n/* harmony export */ });\n/* harmony import */ var _avo_entity__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/entity */ \"./src/avo/entity/index.js\");\n/* harmony import */ var _avo_constants__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/constants */ \"./src/avo/constants.js\");\nfunction _typeof(obj) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (obj) { return typeof obj; } : function (obj) { return obj && \"function\" == typeof Symbol && obj.constructor === Symbol && obj !== Symbol.prototype ? \"symbol\" : typeof obj; }, _typeof(obj); }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\n\nfunction _get() { if (typeof Reflect !== \"undefined\" && Reflect.get) { _get = Reflect.get; } else { _get = function _get(target, property, receiver) { var base = _superPropBase(target, property); if (!base) return; var desc = Object.getOwnPropertyDescriptor(base, property); if (desc.get) { return desc.get.call(arguments.length < 3 ? target : receiver); } return desc.value; }; } return _get.apply(this, arguments); }\n\nfunction _superPropBase(object, property) { while (!Object.prototype.hasOwnProperty.call(object, property)) { object = _getPrototypeOf(object); if (object === null) break; } return object; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); Object.defineProperty(subClass, \"prototype\", { writable: false }); if (superClass) _setPrototypeOf(subClass, superClass); }\n\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\n\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\n\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } else if (call !== void 0) { throw new TypeError(\"Derived constructors may only return object or undefined\"); } return _assertThisInitialized(self); }\n\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\n\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Boolean.prototype.valueOf.call(Reflect.construct(Boolean, [], function () {})); return true; } catch (e) { return false; } }\n\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\n\n\nvar Ground = /*#__PURE__*/function (_Entity) {\n  _inherits(Ground, _Entity);\n\n  var _super = _createSuper(Ground);\n\n  function Ground(app) {\n    var _this;\n\n    var col = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : 0;\n    var row = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : 0;\n    var width = arguments.length > 3 && arguments[3] !== undefined ? arguments[3] : 1;\n    var height = arguments.length > 4 && arguments[4] !== undefined ? arguments[4] : 1;\n    var cutCorner = arguments.length > 5 && arguments[5] !== undefined ? arguments[5] : false;\n\n    _classCallCheck(this, Ground);\n\n    _this = _super.call(this, app);\n    _this._type = 'wall';\n    _this.colour = '#888';\n    _this.solid = true;\n    _this.movable = false;\n    _this.x = col * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE;\n    _this.y = row * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE;\n    _this.shape = _avo_constants__WEBPACK_IMPORTED_MODULE_1__.SHAPES.POLYGON;\n    _this.shapePolygonPath = [];\n    if (cutCorner !== 'nw') _this.shapePolygonPath.push(0, 0);\n    if (cutCorner !== 'ne') _this.shapePolygonPath.push(width * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE, 0);\n    if (cutCorner !== 'se') _this.shapePolygonPath.push(width * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE, height * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE);\n    if (cutCorner !== 'sw') _this.shapePolygonPath.push(0, height * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE);\n    return _this;\n  }\n\n  _createClass(Ground, [{\n    key: \"onCollision\",\n    value: function onCollision(target, collisionCorrection) {\n      _get(_getPrototypeOf(Ground.prototype), \"onCollision\", this).call(this, target, collisionCorrection);\n\n      var app = this._app;\n      var controls = app.rules['cny2023-controls'];\n      var goals = app.rules['cny2023-goals'];\n      if (!controls || !goals) return;\n      if (target === app.hero) goals.triggerLoseScreen();\n    }\n  }]);\n\n  return Ground;\n}(_avo_entity__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n\n\n\n//# sourceURL=webpack://cny2023/./src/avo/entity/types/cny2023/ground.js?");
 
 /***/ }),
 
-/***/ "./src/avo/entity/types/rabbit.js":
-/*!****************************************!*\
-  !*** ./src/avo/entity/types/rabbit.js ***!
-  \****************************************/
+/***/ "./src/avo/entity/types/cny2023/rabbit.js":
+/*!************************************************!*\
+  !*** ./src/avo/entity/types/cny2023/rabbit.js ***!
+  \************************************************/
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ Rabbit)\n/* harmony export */ });\n/* harmony import */ var _avo_entity__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/entity */ \"./src/avo/entity/index.js\");\n/* harmony import */ var _avo_constants__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/constants */ \"./src/avo/constants.js\");\nfunction _typeof(obj) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (obj) { return typeof obj; } : function (obj) { return obj && \"function\" == typeof Symbol && obj.constructor === Symbol && obj !== Symbol.prototype ? \"symbol\" : typeof obj; }, _typeof(obj); }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\n\nfunction _get() { if (typeof Reflect !== \"undefined\" && Reflect.get) { _get = Reflect.get; } else { _get = function _get(target, property, receiver) { var base = _superPropBase(target, property); if (!base) return; var desc = Object.getOwnPropertyDescriptor(base, property); if (desc.get) { return desc.get.call(arguments.length < 3 ? target : receiver); } return desc.value; }; } return _get.apply(this, arguments); }\n\nfunction _superPropBase(object, property) { while (!Object.prototype.hasOwnProperty.call(object, property)) { object = _getPrototypeOf(object); if (object === null) break; } return object; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); Object.defineProperty(subClass, \"prototype\", { writable: false }); if (superClass) _setPrototypeOf(subClass, superClass); }\n\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\n\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\n\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } else if (call !== void 0) { throw new TypeError(\"Derived constructors may only return object or undefined\"); } return _assertThisInitialized(self); }\n\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\n\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Boolean.prototype.valueOf.call(Reflect.construct(Boolean, [], function () {})); return true; } catch (e) { return false; } }\n\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\n\n\nvar Rabbit = /*#__PURE__*/function (_Entity) {\n  _inherits(Rabbit, _Entity);\n\n  var _super = _createSuper(Rabbit);\n\n  function Rabbit(app) {\n    var _this;\n\n    var col = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : 0;\n    var row = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : 0;\n\n    _classCallCheck(this, Rabbit);\n\n    _this = _super.call(this, app);\n    _this._type = 'rabbit';\n    _this.colour = '#c06060';\n    _this.x = col * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE + _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE / 2;\n    _this.y = row * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE + _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE / 2;\n    return _this;\n  }\n  /*\r\n  Section: General Logic\r\n  ----------------------------------------------------------------------------\r\n   */\n\n\n  _createClass(Rabbit, [{\n    key: \"play\",\n    value: function play(timeStep) {\n      var app = this._app;\n\n      _get(_getPrototypeOf(Rabbit.prototype), \"play\", this).call(this, timeStep);\n    }\n  }, {\n    key: \"paint\",\n    value: function paint() {\n      var layer = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : 0;\n      var app = this._app;\n\n      _get(_getPrototypeOf(Rabbit.prototype), \"paint\", this).call(this, layer);\n    }\n    /*\r\n    Section: Animation\r\n    ----------------------------------------------------------------------------\r\n     */\n\n  }, {\n    key: \"getSpriteCol\",\n    value: function getSpriteCol() {}\n  }, {\n    key: \"getSpriteRow\",\n    value: function getSpriteRow() {}\n  }]);\n\n  return Rabbit;\n}(_avo_entity__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n\n\n\n//# sourceURL=webpack://cny2023/./src/avo/entity/types/rabbit.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ Rabbit)\n/* harmony export */ });\n/* harmony import */ var _avo_entity__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/entity */ \"./src/avo/entity/index.js\");\n/* harmony import */ var _avo_constants__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/constants */ \"./src/avo/constants.js\");\nfunction _typeof(obj) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (obj) { return typeof obj; } : function (obj) { return obj && \"function\" == typeof Symbol && obj.constructor === Symbol && obj !== Symbol.prototype ? \"symbol\" : typeof obj; }, _typeof(obj); }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\n\nfunction _get() { if (typeof Reflect !== \"undefined\" && Reflect.get) { _get = Reflect.get; } else { _get = function _get(target, property, receiver) { var base = _superPropBase(target, property); if (!base) return; var desc = Object.getOwnPropertyDescriptor(base, property); if (desc.get) { return desc.get.call(arguments.length < 3 ? target : receiver); } return desc.value; }; } return _get.apply(this, arguments); }\n\nfunction _superPropBase(object, property) { while (!Object.prototype.hasOwnProperty.call(object, property)) { object = _getPrototypeOf(object); if (object === null) break; } return object; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); Object.defineProperty(subClass, \"prototype\", { writable: false }); if (superClass) _setPrototypeOf(subClass, superClass); }\n\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\n\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\n\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } else if (call !== void 0) { throw new TypeError(\"Derived constructors may only return object or undefined\"); } return _assertThisInitialized(self); }\n\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\n\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Boolean.prototype.valueOf.call(Reflect.construct(Boolean, [], function () {})); return true; } catch (e) { return false; } }\n\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\n\n\nvar Rabbit = /*#__PURE__*/function (_Entity) {\n  _inherits(Rabbit, _Entity);\n\n  var _super = _createSuper(Rabbit);\n\n  function Rabbit(app) {\n    var _this;\n\n    var col = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : 0;\n    var row = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : 0;\n\n    _classCallCheck(this, Rabbit);\n\n    _this = _super.call(this, app);\n    _this._type = 'rabbit';\n    _this.colour = '#c06060';\n    _this.x = col * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE + _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE / 2;\n    _this.y = row * _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE + _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE / 2;\n    return _this;\n  }\n  /*\r\n  Section: General Logic\r\n  ----------------------------------------------------------------------------\r\n   */\n\n\n  _createClass(Rabbit, [{\n    key: \"play\",\n    value: function play(timeStep) {\n      var app = this._app;\n\n      _get(_getPrototypeOf(Rabbit.prototype), \"play\", this).call(this, timeStep);\n    }\n  }, {\n    key: \"paint\",\n    value: function paint() {\n      var layer = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : 0;\n      var app = this._app;\n\n      _get(_getPrototypeOf(Rabbit.prototype), \"paint\", this).call(this, layer);\n    }\n    /*\r\n    Section: Animation\r\n    ----------------------------------------------------------------------------\r\n     */\n\n  }, {\n    key: \"getSpriteCol\",\n    value: function getSpriteCol() {}\n  }, {\n    key: \"getSpriteRow\",\n    value: function getSpriteRow() {}\n  }]);\n\n  return Rabbit;\n}(_avo_entity__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n\n\n\n//# sourceURL=webpack://cny2023/./src/avo/entity/types/cny2023/rabbit.js?");
 
 /***/ }),
 
@@ -136,7 +136,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
   \***************************/
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"CNY2023_CEILING_Y\": () => (/* binding */ CNY2023_CEILING_Y),\n/* harmony export */   \"CNY2023_FLOOR_Y\": () => (/* binding */ CNY2023_FLOOR_Y),\n/* harmony export */   \"default\": () => (/* binding */ Levels)\n/* harmony export */ });\n/* harmony import */ var _avo_constants__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/constants */ \"./src/avo/constants.js\");\n/* harmony import */ var _avo_entity_types_rabbit__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/entity/types/rabbit */ \"./src/avo/entity/types/rabbit.js\");\n/* harmony import */ var _avo_entity_types_ground__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! @avo/entity/types/ground */ \"./src/avo/entity/types/ground.js\");\n/* harmony import */ var _avo_entity_types_boost_pad__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! @avo/entity/types/boost-pad */ \"./src/avo/entity/types/boost-pad.js\");\n/* harmony import */ var _avo_rule_types_cny2023_controls__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! @avo/rule/types/cny2023-controls */ \"./src/avo/rule/types/cny2023-controls.js\");\n/* harmony import */ var _avo_rule_types_cny2023_goals__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! @avo/rule/types/cny2023-goals */ \"./src/avo/rule/types/cny2023-goals.js\");\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\n\n\n // import Wall from '@avo/entity/types/wall'\n// import BouncePad from '@avo/entity/types/bounce-pad'\n\n\n\n\n\nvar MIN_X = 0;\nvar MAX_X = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.CNY2023_ROWS * _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE;\nvar ROWS_BETWEEN_BOOSTSPADS = 8;\nvar CNY2023_CEILING_Y = -100 * _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE;\nvar CNY2023_FLOOR_Y = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.CNY2023_ROWS * _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE;\n\nvar Levels = /*#__PURE__*/function () {\n  function Levels(app) {\n    _classCallCheck(this, Levels);\n\n    this._app = app;\n    this.current = 0;\n  }\n\n  _createClass(Levels, [{\n    key: \"reset\",\n    value: function reset() {\n      var app = this._app;\n      app.hero = undefined;\n      app.entities = [];\n      app.clearRules();\n      app.camera.target = null;\n      app.camera.x = 0;\n      app.camera.y = 0;\n      app.camera.zoom = 1;\n      app.playerAction = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.PLAYER_ACTIONS.IDLE;\n      app.setInteractionMenu(false);\n    }\n  }, {\n    key: \"load\",\n    value: function load() {\n      var level = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : 0;\n      var app = this._app;\n      this.current = level;\n      this.reset();\n      this.generate_default();\n    }\n  }, {\n    key: \"reload\",\n    value: function reload() {\n      this.load(this.current);\n    }\n    /*\r\n    Default level.\r\n     */\n\n  }, {\n    key: \"generate_default\",\n    value: function generate_default() {\n      var app = this._app;\n      app.hero = app.addEntity(new _avo_entity_types_rabbit__WEBPACK_IMPORTED_MODULE_1__[\"default\"](app, (_avo_constants__WEBPACK_IMPORTED_MODULE_0__.CNY2023_COLS - 1) / 2, 1));\n      app.camera.x = 0;\n      app.camera.y = 0; //app.camera.target = app.hero\n\n      app.addRule(new _avo_rule_types_cny2023_controls__WEBPACK_IMPORTED_MODULE_4__[\"default\"](app));\n      app.addRule(new _avo_rule_types_cny2023_goals__WEBPACK_IMPORTED_MODULE_5__[\"default\"](app)); // app.addEntity(new Wall(app, 0, 0, 1, 16))  // West Wall\n      // app.addEntity(new Wall(app, 23, 0, 1, 16))  // East Wall\n      // app.addEntity(new Wall(app, 1, 0, 22, 1))  // North Wall\n      // app.addEntity(new Wall(app, 0, CNY2023_ROWS - 1, CNY2023_COLS, 1))  // South Wall\n\n      app.addEntity(new _avo_entity_types_ground__WEBPACK_IMPORTED_MODULE_2__[\"default\"](app, 0, _avo_constants__WEBPACK_IMPORTED_MODULE_0__.CNY2023_ROWS - 1, _avo_constants__WEBPACK_IMPORTED_MODULE_0__.CNY2023_COLS, 1)); // South Wall\n\n      var boostPadWidth = 10;\n      app.addEntity(new _avo_entity_types_boost_pad__WEBPACK_IMPORTED_MODULE_3__[\"default\"](app, (_avo_constants__WEBPACK_IMPORTED_MODULE_0__.CNY2023_COLS - boostPadWidth) / 2, _avo_constants__WEBPACK_IMPORTED_MODULE_0__.CNY2023_ROWS - 2, boostPadWidth, 1)); // Boostpad\n\n      for (var y = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE * 4; y >= CNY2023_CEILING_Y; y -= _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE * ROWS_BETWEEN_BOOSTSPADS) {\n        this.createBoucePad(y);\n        console.log(y);\n      }\n    }\n  }, {\n    key: \"createBoucePad\",\n    value: function createBoucePad() {\n      var y = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : 0;\n      var app = this._app;\n      var width = Math.random() * 8 + 2;\n      var height = 1;\n      var x = Math.random() * (MAX_X - MIN_X - width);\n      var row = y / _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE;\n      var col = x / _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE;\n      app.addEntity(new _avo_entity_types_boost_pad__WEBPACK_IMPORTED_MODULE_3__[\"default\"](app, col, row, width, height));\n    }\n  }]);\n\n  return Levels;\n}();\n\n\n\n//# sourceURL=webpack://cny2023/./src/avo/levels.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"CNY2023_CEILING_Y\": () => (/* binding */ CNY2023_CEILING_Y),\n/* harmony export */   \"CNY2023_FLOOR_Y\": () => (/* binding */ CNY2023_FLOOR_Y),\n/* harmony export */   \"default\": () => (/* binding */ Levels)\n/* harmony export */ });\n/* harmony import */ var _avo_constants__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/constants */ \"./src/avo/constants.js\");\n/* harmony import */ var _avo_entity_types_cny2023_rabbit__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/entity/types/cny2023/rabbit */ \"./src/avo/entity/types/cny2023/rabbit.js\");\n/* harmony import */ var _avo_entity_types_cny2023_ground__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! @avo/entity/types/cny2023/ground */ \"./src/avo/entity/types/cny2023/ground.js\");\n/* harmony import */ var _avo_entity_types_cny2023_boost_pad__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! @avo/entity/types/cny2023/boost-pad */ \"./src/avo/entity/types/cny2023/boost-pad.js\");\n/* harmony import */ var _avo_rule_types_cny2023_controls__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! @avo/rule/types/cny2023-controls */ \"./src/avo/rule/types/cny2023-controls.js\");\n/* harmony import */ var _avo_rule_types_cny2023_goals__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! @avo/rule/types/cny2023-goals */ \"./src/avo/rule/types/cny2023-goals.js\");\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\n\n\n\n\n\n\n\nvar MIN_X = 0;\nvar MAX_X = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.CNY2023_ROWS * _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE;\nvar ROWS_BETWEEN_BOOSTSPADS = 8;\nvar CNY2023_CEILING_Y = -100 * _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE;\nvar CNY2023_FLOOR_Y = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.CNY2023_ROWS * _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE;\n\nvar Levels = /*#__PURE__*/function () {\n  function Levels(app) {\n    _classCallCheck(this, Levels);\n\n    this._app = app;\n    this.current = 0;\n  }\n\n  _createClass(Levels, [{\n    key: \"reset\",\n    value: function reset() {\n      var app = this._app;\n      app.hero = undefined;\n      app.entities = [];\n      app.clearRules();\n      app.camera.target = null;\n      app.camera.x = 0;\n      app.camera.y = 0;\n      app.camera.zoom = 1;\n      app.playerAction = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.PLAYER_ACTIONS.IDLE;\n      app.setInteractionMenu(false);\n    }\n  }, {\n    key: \"load\",\n    value: function load() {\n      var level = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : 0;\n      var app = this._app;\n      this.current = level;\n      this.reset();\n      this.generate_default();\n    }\n  }, {\n    key: \"reload\",\n    value: function reload() {\n      this.load(this.current);\n    }\n    /*\r\n    Default level.\r\n     */\n\n  }, {\n    key: \"generate_default\",\n    value: function generate_default() {\n      var app = this._app;\n      app.hero = app.addEntity(new _avo_entity_types_cny2023_rabbit__WEBPACK_IMPORTED_MODULE_1__[\"default\"](app, (_avo_constants__WEBPACK_IMPORTED_MODULE_0__.CNY2023_COLS - 1) / 2, 1));\n      app.camera.x = 0;\n      app.camera.y = 0; //app.camera.target = app.hero\n\n      app.addRule(new _avo_rule_types_cny2023_controls__WEBPACK_IMPORTED_MODULE_4__[\"default\"](app));\n      app.addRule(new _avo_rule_types_cny2023_goals__WEBPACK_IMPORTED_MODULE_5__[\"default\"](app)); // app.addEntity(new Wall(app, 0, 0, 1, 16))  // West Wall\n      // app.addEntity(new Wall(app, 23, 0, 1, 16))  // East Wall\n      // app.addEntity(new Wall(app, 1, 0, 22, 1))  // North Wall\n      // app.addEntity(new Wall(app, 0, CNY2023_ROWS - 1, CNY2023_COLS, 1))  // South Wall\n\n      app.addEntity(new _avo_entity_types_cny2023_ground__WEBPACK_IMPORTED_MODULE_2__[\"default\"](app, 0, _avo_constants__WEBPACK_IMPORTED_MODULE_0__.CNY2023_ROWS - 1, _avo_constants__WEBPACK_IMPORTED_MODULE_0__.CNY2023_COLS, 1)); // South Wall\n\n      var boostPadWidth = 10;\n      app.addEntity(new _avo_entity_types_cny2023_boost_pad__WEBPACK_IMPORTED_MODULE_3__[\"default\"](app, (_avo_constants__WEBPACK_IMPORTED_MODULE_0__.CNY2023_COLS - boostPadWidth) / 2, _avo_constants__WEBPACK_IMPORTED_MODULE_0__.CNY2023_ROWS - 2, boostPadWidth, 1)); // Boostpad\n\n      for (var y = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE * 4; y >= CNY2023_CEILING_Y; y -= _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE * ROWS_BETWEEN_BOOSTSPADS) {\n        this.createBoucePad(y);\n        console.log(y);\n      }\n    }\n  }, {\n    key: \"createBoucePad\",\n    value: function createBoucePad() {\n      var y = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : 0;\n      var app = this._app;\n      var width = Math.random() * 8 + 2;\n      var height = 1;\n      var x = Math.random() * (MAX_X - MIN_X - width);\n      var row = y / _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE;\n      var col = x / _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE;\n      app.addEntity(new _avo_entity_types_cny2023_boost_pad__WEBPACK_IMPORTED_MODULE_3__[\"default\"](app, col, row, width, height));\n    }\n  }]);\n\n  return Levels;\n}();\n\n\n\n//# sourceURL=webpack://cny2023/./src/avo/levels.js?");
 
 /***/ }),
