+++ src/avo/entity/types/cny2023/boost-pad.js
