+++ src/avo/entity/types/cny2023/bounce-pad.js
