+++ src/avo/entity/types/cny2023/rabbit.js
